chore(notification): plugged health check into the handler
