+++ services/notification/cmd/server/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"log"
 	"net/http"
 	"os"
@@ -16,31 +15,6 @@ import (
 	api "github.com/vantavoids/ft_transcendence/services/notification/internal/transport/api"
 	broker "github.com/vantavoids/ft_transcendence/services/notification/internal/transport/broker"
 )
-
-type healthCheck struct {
-	Name        string  `json:"name"`
-	Status      string  `json:"status"`
-	Description *string `json:"description"`
-}
-
-type healthReport struct {
-	Status string        `json:"status"`
-	Checks []healthCheck `json:"checks"`
-}
-
-// healthz reports the service's liveness. notification has no
-// app-layer DB client wired yet, so the only registered check is its own
-// liveness; the backing DB is gated separately by the compose healthcheck.
-func healthz(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	json.NewEncoder(w).Encode(healthReport{
-		Status: "Healthy",
-		Checks: []healthCheck{
-			{Name: "self", Status: "Healthy"},
-		},
-	})
-}
 
 func main() {
 	ctx := context.Background()
